Add an implementation of sound environment sampler
A direct occlusion is determined by testing traces from
an emitter to a listener.

Also, two mutually exclusive effects are applied
based on contents of an emitter:

* An underwater flanger

* A reverberation effect

An environment is sampled by testing a solid world
against randomly directed rays outgoing from an emitter origin.
If a ray hits a solid, a secondary emission is performed
by testing a trace in the solid world to the listener for
determining reflected sound occlusion.

AL standart reverberation parameters are extracted from the
tracing results and saved for interpolation during
environment updates of an emitter.

A priority queue of updates is used, newly added or important
sounds have higher priority. Low-priority updates might be
deferred to next sources update frame.

An environment sampler works in the sound background thread.
In order to achieve thread safety the client/listen server
collision model instance is cloned if needed and sound
library calls use the cloned collision model.

Issues: there is no collision model in demo playing, so
sound effects are not reproducible during demo playback.