Lower Shockwave projectile speed/add switch delay
This change is needed not only because it seems overpowered,
but also because it has to be distinct in aiming from RL,
otherwise human brain gets confused and aiming becomes worse
for all these weapons.

The second change is intended for testing weapon switch delay feel
(not to mention LG/SW, PG/SW combos are extemely powerful and
lead to be abused producing very annoying play style).