Test shapes bounding spheres against a trace line
Precompute bounding spheres for brushes, patches
and patch facets and use center-to-trace-line distance
test for quick collision shape rejection
in CM_CollideBox(). Assume if the distance is greater
than sum of shape bounding sphere radius and traced
box bounding sphere radius, a shape can be safely
rejected. Absolute trace mins/maxs testing against
a shape bounds is still useful and serves not only
for fast coarse shapes rejection, but for rejection of
shapes that cannot intersect the actual line segment
(thats to avoid testing of center-to-segment distance).

Add a separate code path using the newly introduced
optimizations and use it for line tracing (not position
testing) in solid world.

Keep the old code path for all other cases since
this optimization is not applicable there.

TODO: Actually vectorize the implementation.